fix(ci): keep connection payload helpers off hotspots
